+++ src/navigation/EventNavigator.tsx
@@ -87,20 +87,4 @@ export const EventNavigator = () => {
     content = <ErrorPage error={error} />;
   }
   return <>{content}</>;
-
-  // return (
-  //   <Tab.Navigator
-  //     screenOptions={() => ({
-  //       tabBarIndicatorStyle: {
-  //         backgroundColor: "black", //Indicatorの色
-  //         height: 1.5
-  //       }
-  //     })}
-  //   >
-  //     <Tab.Screen name="全体" component={AllEventList} />
-  //     <Tab.Screen name="お知らせ1" component={EventList1} />
-  //     <Tab.Screen name="お知らせ2" component={EventList2} />
-  //     <Tab.Screen name="お知らせ3" component={EventList3} />
-  //   </Tab.Navigator>
-  // );
 };
