+++ src/pages/searchPage/FavList.tsx
@@ -1,32 +1,17 @@
-import { useState } from "react";
 import { View, Image } from "native-base";
-import { FlatList, RefreshControl } from "react-native";
+import { FlatList } from "react-native";
 import {
   SPACING_SQUARE,
   SPACING_SQUARE_SIZE
 } from "@components/styles/theme/layout";
 
 //redux==============================
-import { useAppDispatch, useAppSelector } from "@Redux/hook";
+import { useAppSelector } from "@Redux/hook";
 import { selectLikedPost } from "@Redux/postsSlice";
-import { fetchAllPosts, updateLike } from "@Redux/postActions";
-export const FavList = ({ route }) => {
-  const dispatch = useAppDispatch();
+
+export const FavList = () => {
   const _favs = useAppSelector((state) => selectLikedPost(state));
 
-  const [refreshing, setRefreshing] = useState(false);
-
-  const _onRefresh = () => {
-    setRefreshing(true);
-    // dispatch(
-    //   updateLike({
-    //     postId: item.postId,
-    //     isLiked: trigger
-    //   } as Post)
-    // );
-    dispatch(fetchAllPosts());
-    setRefreshing(false);
-  };
   const _renderItem = ({ item }) => {
     return (
       <View>
@@ -37,7 +22,7 @@ export const FavList = ({ route }) => {
             margin: SPACING_SQUARE
           }}
           source={{ uri: item.postedImage }}
-          resizeMode="contain"
+          resizeMode="cover"
           alt="post"
         />
       </View>
@@ -50,10 +35,6 @@ export const FavList = ({ route }) => {
         keyExtractor={(_, index) => index.toString()}
         renderItem={_renderItem}
         numColumns={2}
-        // refreshing={refreshing}
-        // refreshControl={
-        //   <RefreshControl refreshing={refreshing} onRefresh={_onRefresh} />
-        // }
       />
     </View>
   );
