--- src/pages/searchPage/SearchLists.tsx
@@ -1,11 +1,0 @@
-// //イベントcategory1
-
-// import SearchList from "@components/SearchList";
-
-// // export const GenrePickList = ({}) => {
-// //   return <GenreList />;
-// // };
-
-// export const ProductList = ({}) => {
-//   return <SearchList />;
-// };
